Added the registration information.
Users may now register their device with a group. Registration with a group dictates which set of apps will be returned to you. The value in the registration page is persisted throughout the app meaning that updates will use that same registration value. Whenever a user checks for updates, that registration value is sent.