Added the capability to build from the command line.
Android's auto-generated build tools have been added to allow the APK to be built from the command line. Further, the keystore has been added, so a release build can be built as well.

Note: In order to build from the command line, be sure to set the ANDROID_HOME environment variable to the location of the Android SDK home.